+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 
 ### Fixed
 
--  Fix bug where which window was on top was confusing
--  Fix bug where ranges that should not be visible are visible on startup
+- Fix bug where which window was on top was confusing
+- Fix bug where ranges that should not be visible are visible on startup and when enabling filter ([Issue #86](https://github.com/nyetwurk/ecuxplot/issues/86))
 
 ## [1.1.4] - 2025-10-29